napi: use V8 instanceof tests for napi_instanceof()
