--- test/addons-abi/test_instanceof/test2.js
@@ -1,4 +1,0 @@
-var addon = require( "./build/Release/test_instanceof" );
-function F() {}
-var x = 1;
-addon.doInstanceOf( x, F );
